move auth forms to components folder
